requirements.txt
## Requirements

### nft_offers.py
chia-blockchain==1.5.1